pyproject.toml
[project]
name = "lbg-ipi-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.22.0",
    "google-generativeai>=0.8.6",
    "litellm>=1.80.13",
    "pandas>=2.3.3",
    "postal>=1.1.11",
    "psutil>=7.2.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
